l54: isr80h command table linked + registered (Lecture 54)
PeachOS64BitCourse commit 2c4505a ported with three SamOs-side
width-bug fixes:

Makefile / build.sh:
  isr80h.o, io.o, heap.o, misc.o, process.o added to FILES
  with build rules. mkdir build/isr80h.

src/kernel.c:
  task/task.h, task/process.h, isr80h/isr80h.h includes added.
  After gdt_set_tss in kernel_main:
    isr80h_register_commands();
  Populates the static isr80h_commands[] table that
  isr80h_handler dispatches into on int 0x80.

SamOs deviations - 64-bit-width cast fixes the 32-bit source
got away with:

src/isr80h/io.c:
  isr80h_command2_getkey now returns (void*)((intptr_t)c)
    instead of (void*)((int)c). The 32-bit form truncated to
    int (same size as void* on 32-bit) then cast back; on
    x86_64 the int-to-pointer cast diagnostic fires because
    sizes differ.
  isr80h_command3_putchar reads its char arg as
    (char)(intptr_t)task_get_stack_item(...) instead of
    (char)(int)... - stack item is void* (8 bytes); intermediate
    needs to be pointer-width.

src/isr80h/heap.c:
  isr80h_command4_malloc reads its size_t arg as
    (size_t)(uintptr_t)task_get_stack_item(...) instead of
    (int)... - preserves sizes > 2 GiB and silences the
    pointer-to-int-of-different-size diagnostic.

src/isr80h/misc.c:
  isr80h_command0_sum reads both int operands through
    (int)(intptr_t)task_get_stack_item(...). Return wraps the
    sum in (void*)(intptr_t).

The cast bugs are correctness issues, not just warnings - on
32-bit the (int) cast was a no-op bit reinterpretation; on
x86_64 it would truncate the high 32 bits of an address. With
-Werror, GCC refuses to compile until they are fixed.

Nothing in kernel_main sends an int 0x80 yet, so the command
table is dormant. The unit of test is "the link works and
isr80h_register_commands does not fault".

Test: tests64/L54-isr80h-linked.sh - same tokens as L52/L53
(tss ready).

Verified on VGA:
  Hello 64-bit!
  e820 total: 267910144
  ABCmultiheap ready
  tss ready

Doc: docs/64bit/chapters/L54-isr80h-linked.md walks the
cast-bug pattern and explains why isr80h is dormant at L54
(no user process; interrupts off).

Next: Lecture 55+ - load and launch a user process so the
isr80h table actually gets exercised.

+++ src/isr80h/heap.c
@@ -4,7 +4,11 @@
 #include <stddef.h>
 
 void* isr80h_command4_malloc(struct interrupt_frame* frame){
-    size_t size = (int)task_get_stack_item(task_current(), 0);
+    // L54 - widen the cast: stack items are void* (8 bytes on
+    // x86_64), and the user is asking for a size_t. Going
+    // through intptr_t avoids the pointer-to-int-of-different-
+    // size diagnostic and preserves any size > 2 GiB.
+    size_t size = (size_t)(uintptr_t)task_get_stack_item(task_current(), 0);
     return process_malloc(task_current()->process, size);
 }
